Fix incorrect import in grafana/ui
