+++ packages/grafana-ui/src/utils/index.ts
@@ -6,6 +6,7 @@ export * from './fieldDisplay';
 export * from './validate';
 export { getFlotPairs } from './flotPairs';
 export * from './slate';
+export * from './dataLinks';
 
 // Export with a namespace
 import * as DOMUtil from './dom'; // includes Element.closest polyfil
